+++ Year2024/day11/utils.h
@@ -1,13 +1,14 @@
 #ifndef UTILS_H
 #define UTILS_H
 
+#include <cstdint>
 #include <map>
 #include <string>
 #include <vector>
 
 using namespace std;
 
-map<string, int> evol(const map<string, int>& termites, const map<string, vector<string>>& evolution);
-int runFullEvolution(const int& iter, const map<string, vector<string>>& evolution, const string& starting);
+map<string, uint64_t> evol(const map<string, uint64_t>& termites, const map<string, vector<string>>& evolution);
+uint64_t runFullEvolution(const int& iter, const map<string, vector<string>>& evolution, const string& starting);
 
 #endif
